Introduce "complete conformance" to conformance level section
Here's a suggestion of how I'm thinking we can resolve the issue I raised on AG on Tuesday. I think it is important to keep a clear distinction between what the requirements are, and what the margin of error is that we're allowing on it. Put in other words, it needs to be clear that every image missing an alt attribute is an accessibility issue; That it isn't OK for content authors to intentionally pick images that are difficult to describe and leave the alt attribute off.

This definitely needs more work, but it shows the direction I'm thinking for a solution. I do think it is important that we address this in FPWD, even if it is only a partial solution, some acknowledgement of the problem is important in my opinion.