Remove old code, replace with link to repo, etc.
It came to my attention that people are still using this page via GitHub.io, but since all of this information was deleted from the FPWD, it is important that this page be current to prevent misunderstandings. This commit deletes the old obsolete sampleJS code (under "code samples"), corrects some important values in the plain language tutorial and pseudocode (under the resources tab), and provides links to the GitHub repo for the authoritative APCA code and methods.