CORRECT OBSOLETE CODE THAT IS CAUSING PROBLEMS
Okay, in lieu of pull request  #213, this pull request addresses Jeanne's concerns and also get the bod information away from any links.

TO BE CLEAR: the document this pull request is correcting **was never supposed to be public facing and linked** per my understanding  — but the correct code was deleted from the FPWD,  yet in this doc was BAD OBSOLETE CODE that should have been replaced by the FPWD. This document is STILL LINKED, and now there are several instances of bade code in the wild as a result.

For reasons that are unclear, the links to this BAD CODE keep popping up as if someone is attempting to sabotage this project.

PLEASE PUSH THIS PULL REQUEST THROUGH. This supersedes the previous request (#213)

In THIS request, I left the code and only find the things that would have been fixed if the correct code was not deleted from the FPWD.