Mark deprecated link
NOTE: this material is old and obsolete, yet it is still being linked to and appearing in google searches. It need to be marked as DEPRECATED with a link to the current at the very least.

THe old link: https://w3c.github.io/silver/guidelines/methods/Method-font-characteristic-contrast.html is still active.

Can we PLEASE pretty please with sugar or honey or whatnot on top PLEASE apply this PR.

All of the OTHER guideline assets have been removed BUT NOT THIS ONE. Why?