Visual Contrast Methods
Repurposed "examples" into "Test Techniques", and placed explanatory materials there.

Updated Lookup tables to current versions

Updated "tests" page for clarity and brevity, and also to encompass the new testing concepts of "objective" and  "Qualified" tests, etc.

Added links to resources, both W3 and external, and added the beginning of a bibliography.

Note: This should not be called "font characteristics"  what is the pricess of correcting the name?